requirements.txt
acachecontrol==0.3.5
aenum==3.1.11
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cloudinary==1.40.0
constantly==23.10.4
cryptography==42.0.7
decorator==5.1.1
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-oauth-toolkit==2.4.0
django-rest-passwordreset==1.4.1
djangorestframework==3.15.1
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
ecdsa==0.19.0
firebase-admin==6.5.0
flatdict==4.0.1
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jwcrypto==1.5.6
msgpack==1.0.8
multidict==6.0.5
oauthlib==3.2.2
packaging==24.0
pika==1.3.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydash==8.0.1
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.5
requests==2.32.2
retry2==0.9.5
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
xmltodict==0.13.0
yarl==1.9.4
zope.interface==6.4.post2
